Update index.html
Added a button to the bottom of the page that would bring a user to back
to the top of the page if it was clicked.  I also added a image that
owuld go over the front cover image that when hovered over it would let
the user know to scroll down just in case that thought the background
image was the whole page.