update: index.html
I made some changes so the hover features change on mobile device screen
sizes.  Now mobile users will be able to use the website the same as on
desktop.  I also added in a menu icon that appears when in mobile.  It
can be used to hide and show the menu.